refactor(FNP-55): remove unused SVG files
